clip_extracting/requirements.txt
bitarray==3.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
cvcuda-cu11==0.14.0
filelock==3.13.1
fsspec==2024.6.1
idna==3.4
Jinja2==3.1.4
Mako==1.3.10
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==11.0.0
platformdirs==4.3.8
pycuda==2025.1
PyNvVideoCodec==1.0.2
pytools==2025.1.6
requests==2.28.1
siphash24==1.7
sympy==1.13.3
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.67.1
triton==2.1.0
typing_extensions==4.12.2
urllib3==1.26.13
